Refactor PBR routine to be more replaceable
